refactor Added index filed which exports everything and started cleaninig up tools file
