+++ app/src/components/market/routes/market_routes.tsx
@@ -7,7 +7,7 @@ import { useCheckContractExists, useMarketMakerData } from '../../../hooks'
 import { useConnectedWeb3Context } from '../../../hooks/connectedWeb3'
 import { MarketDetailsPage } from '../../../pages'
 import { getLogger } from '../../../util/logger'
-import { isAddress } from '../../../util/tools'
+import { isAddress } from '../../../util/tools/base'
 import { ThreeBoxComments } from '../../comments'
 import { InlineLoading } from '../../loading'
 import { MarketNotFound } from '../sections/market_not_found'
