+++ app/src/components/modal/modal_deposit_withdraw/index.tsx
@@ -11,7 +11,7 @@ import { useConnectedWeb3Context } from '../../../hooks'
 import { ERC20Service, XdaiService } from '../../../services'
 import { bridgeTokensList, getNativeAsset, getToken, networkIds } from '../../../util/networks'
 import { getImageUrl } from '../../../util/token'
-import { formatBigNumber, formatNumber, waitForConfirmations } from '../../../util/tools'
+import { formatBigNumber, formatNumber, waitForConfirmations } from '../../../util/tools/base'
 import { ExchangeType, Token, TransactionStep } from '../../../util/types'
 import { Button, ButtonStateful } from '../../button'
 import { ButtonStates } from '../../button/button_stateful'
