+++ app/src/hooks/useBalance.tsx
@@ -7,7 +7,7 @@ import { useTokens } from '../hooks'
 import { XdaiService } from '../services'
 import { getLogger } from '../util/logger'
 import { bridgeTokensList, getNativeAsset, networkIds } from '../util/networks'
-import { formatBigNumber, formatNumber } from '../util/tools'
+import { formatBigNumber, formatNumber } from '../util/tools/base'
 import { KnownTokenValue, Token } from '../util/types'
 
 const logger = getLogger('Hooks::ConnectedBalance')
