+++ app/src/hooks/useBlockchainMarketMakerData.tsx
@@ -5,7 +5,7 @@ import { useCallback, useEffect, useState } from 'react'
 import { ERC20Service, MarketMakerService, OracleService } from '../services'
 import { getLogger } from '../util/logger'
 import { getArbitratorFromAddress } from '../util/networks'
-import { isScalarMarket, promiseProps } from '../util/tools'
+import { isScalarMarket, promiseProps } from '../util/tools/base'
 import { BalanceItem, MarketMakerData, Status, Token } from '../util/types'
 
 import { useConnectedWeb3Context } from './connectedWeb3'
