+++ app/src/util/call_in_chunks.ts
@@ -1,4 +1,4 @@
-import { delay } from './tools'
+import { delay } from './tools/base'
 
 export type Range = [number, number]
