+++ app/src/util/tools/fpmm/liquidity/calcAddFundingSendAmounts.ts
@@ -7,7 +7,14 @@ import { calcAddFundingDepositedAmounts } from './calcAddFundingDepositedAmounts
  * @param addedFunds - the amount of collateral being added to the market maker as liquidity
  * @param poolBalances - the market maker's balances of outcome tokens
  */
-export const calcAddFundingSendAmounts = (addedFunds: BigNumber, poolBalances: BigNumber[]): BigNumber[] => {
+export const calcAddFundingSendAmounts = (
+  addedFunds: BigNumber,
+  poolBalances: BigNumber[],
+  poolShareSupply?: BigNumber,
+): Maybe<BigNumber[]> => {
+  if (poolShareSupply && poolShareSupply.eq(0)) {
+    return null
+  }
   const depositAmounts = calcAddFundingDepositedAmounts(addedFunds, poolBalances)
 
   const sendAmounts = depositAmounts.map(depositAmount => addedFunds.sub(depositAmount))
